+++ controller/login.ts
@@ -1,11 +1,8 @@
-import crypto from "crypto";
 import { Request, Response } from "express";
-import jwt from "jsonwebtoken";
 import { Prisma } from "@prisma/client";
-import config from "@/config/token";
-import db from "@/db";
 import isLogin from "@/utils/login";
-import { ConnectionError, BadRequest, NotFound } from "@/types/error";
+import { ConnectionError, BadRequest } from "@/types/error";
+import { createTokens, verifyUserinfo } from "@/utils/verify";
 
 export default {
   post,
@@ -13,31 +10,15 @@ export default {
 
 async function post(req: Request, res: Response) {
   try {
-    const user_id = await isLogin(req, res);
-    if (user_id) throw new BadRequest("Already logged in");
-    const { username, password: plain } = req.body as Prisma.UserCreateInput;
-    // 유저 정보 확인
-    const where = { username };
-    const select = { id: true, salt: true, password: true };
-    const user = await db.user.findUnique({ where, select });
-    if (!user) throw new NotFound({ username });
-    const { id, salt, password } = user;
+    const is_logged_in = await isLogin(req, res);
+    // 이미 로그인 상태인 경우 에러 발생
+    if (is_logged_in) throw new BadRequest("Already logged in");
+    // 입력값 추출
+    const { username, password } = req.body as Prisma.UserCreateInput;
     // 비밀번호 확인
-    const encoded = crypto
-      .pbkdf2Sync(plain, salt, 100000, 64, "sha512")
-      .toString("base64");
-    // pw가 일치하지 않아도 NotFound 에러 발생
-    // pw가 틀렸다고 하면 username의 존재를 알려주기 때문에 DB 간접적으로 노출
-    if (password !== encoded) throw new NotFound({ username });
+    const id = await verifyUserinfo(username, password);
     // JWT 토큰 생성
-    const access = jwt.sign({ id }, config.ACCESS_TOKEN!, {
-      expiresIn: "1h",
-    });
-    const refresh = jwt.sign({ id }, config.REFRESH_TOKEN!, {
-      expiresIn: "30d",
-    });
-    // DB에 refresh 토큰 저장
-    await db.refresh.create({ data: { user_id: id, refresh, access } });
+    const { access, refresh } = await createTokens(id);
     // 토큰 전송
     res.status(200).json({ access, refresh });
   } catch (error) {
